Add Ability Fields to Card Model & Table
Add the Ability fields to the Card model and table because the Ability
of a card is optional. Ability is one per card if the card has one and
keeps the number of helper tables that need to use with a card to a
minimum, which makes maintaining such a system too much.