Add Pokemon TCG SDK Call as Method to Cards Model
Add Pokemon TCG SDK call as a method to the Cards model to be easier to
stub put then directly and try to learn how the inside of the Pokemon
TCG SDK. However, it means we can switch Pokemon TCG SDK gem if we wish
something different and not break all of the tests. It is assuming that
the new Pokemon SDK gem will return a similar response.