Add Helper Methods From Each Helper Class of Cards
Add helper methods from each helper class of Cards because each card
helper class Attacks, Card Groups, Types, and Texts are responsible for
managing their data. So that if there is a change to how the data is
structured. There is a single place to change that data without changing
different parts of the codebase. Which can lead to hard to manage
systems.