Fix Header Style to Use Info Instead of Secondary
Fix the Header style to use Info instead of the secondary style since
used in the purge back up button as the secondary is redder, so it works
well for control meant to delete everything. Well, the header is
intended to be a contrast to the body so that it stands out from it.