Add selectMaterialUiSelectOption to Jest Helpers
Add the `selectMaterialUiSelectOption` method to Jest Helpers because
the select Material UI uses is react-select under the hood. As a result,
it is not using select form element but div elements, making it hard to
select a part from that input for testing.

`selectMaterialUiSelectOption` finds text you want to use and makes the
selection for you.

---
Ref:

https://github.com/testing-library/react-testing-library/issues/322#issuecomment-581650108