Add Search Form Component For Search Form Elements
Add the Search Form component for Search Form elements so that the
search component itself can be for the toggle part, and the rest can
handle the form. Furthermore, this can allow the form to be placed in
another part of the application as it does not need to hold the toggle
part of the search form.