Add to_h to Types Model, Return Hash Vers of Types
Add the `to_h` method to the Types model to return a hash version of
Types so that other sections of the application does not need to access
the database every time they need to know if they type exist of not.
Furthermore, this change can reduce the number of calls to the database,
thus lowering wait time and communications to the database.